+++ test/mocks/PrecompileCallerMock.sol
@@ -13,7 +13,7 @@ contract PrecompileCallerMock {
         (bool success, bytes memory response) = ASSETS_PRECOMPILE_ADDRESS.call{gas: 216_147}(
             abi.encodeWithSelector(
                 ASSETS_CONTRACT.depositTo.selector,
-                uint16(101),
+                uint32(101),
                 abi.encodePacked(bytes32(bytes20(address(0xdAC17F958D2ee523a2206206994597C13D831ec7)))),
                 abi.encodePacked(bytes32(bytes20(address(0x2)))),
                 amount
